SNE-77 Interior Hero structure
