+++ blocks/hero/hero.css
@@ -1,1 +1,1 @@
-.hero-container .hero-wrapper{max-width:unset;padding:0}.hero{position:relative;padding:40px 24px;min-height:300px}.hero h1{max-width:1200px;margin-left:auto;margin-right:auto;color:var(--background-color)}.hero picture{position:absolute;z-index:-1;inset:0;object-fit:cover;box-sizing:border-box}.hero img{object-fit:cover;width:100%;height:100%}@media(width >= 900px){.hero{padding:40px 32px}}
+.usa-hero{font-family:Roboto, Helvetica, Arial, sans-serif;font-size:1.24rem;line-height:1.5;background-image:url("/icons/hero.jpg");background-position:center;background-size:cover;color:#fff}.usa-hero>.grid-container{position:relative;padding-bottom:2rem;padding-top:2rem}.usa-hero__callout{background-color:#003a4d;padding:2rem}@media all and (min-width: 40em){.usa-hero__callout{max-width:20rem}}.usa-hero__heading{margin-bottom:0;margin-top:0;clear:both;font-family:Montserrat, Roboto, Helvetica, Arial, sans-serif;font-size:1.99rem;line-height:1.2;font-weight:700;color:#00bde3;line-height:1.2}*+.usa-hero__heading{margin-top:1.5em}.usa-hero__heading+*{margin-top:1em}.usa-hero__heading--alt{color:#fff;display:block}
