+++ blocks/hero/hero.scss
@@ -1,38 +1,8 @@
-// stylelint-disable
-.hero-container .hero-wrapper {
-	max-width: unset;
-	padding: 0;
-}
+@use '../../styles/ne-global-imports' as *;
+@use '~uswds/usa-hero/src/styles/usa-hero';
 
-.hero {
-	position: relative;
-	padding: 40px 24px;
-	min-height: 300px;
-}
+$block: 'usa-hero';
 
-.hero h1 {
-	max-width: 1200px;
-	margin-left: auto;
-	margin-right: auto;
-	color: var(--background-color);
-}
-
-.hero picture {
-	position: absolute;
-	z-index: -1;
-	inset: 0;
-	object-fit: cover;
-	box-sizing: border-box;
-}
-
-.hero img {
-	object-fit: cover;
-	width: 100%;
-	height: 100%;
-}
-
-@media (width >= 900px) {
-	.hero {
-		padding: 40px 32px;
-	}
+.#{$block} {
+	// somthing
 }
